LPD-68099 Isolate ScheduleField and related functions to reuse them
